refactor lmstudio
